Name service instance counter instead of redis in state lab
On first glance the app for this lab looks like it is specifically searching for a service instance called 'redis' but it isn't

Instead it is looking in VCAP_SERVICES for the key _starting with_ redis. i.e. on PWS the name of the redis service in the marketplace is 'rediscloud'

Signed-off-by: Colin Simmons <colin.simmons@engineerbetter.com>